implementation of controller and inheritance
