+++ src/components/trip-event-wrapper.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from './abstract-component.js';
 
 const months = [`Jan`,
   `Feb`,
@@ -13,27 +13,13 @@ const months = [`Jan`,
   `Nov`,
   `Dec`];
 
-export default class TripEventWrapper {
+export default class TripEventWrapper extends AbstractComponent {
   constructor({startDate}) {
+    super();
     this._startDate = startDate;
-    this._element = null;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-    return this._element;
   }
 
   getTemplate() {
-
     return `<ul class="trip-days">
             <li class="trip-days__item day">
               <div class="day__info">
